recursionIndent should be thread-local
(But that wouldn't help much when using this in an actual
multi-threaded case anyway, as we don't prefix output lines with any
thread id, and we also probably would cheerfully mix output from
multiple threads onto the same lines. If we really want to use this in
multi-threaded case, we probably want to collect the output for each
output line completely into a (thread-local) buffer before outputting
it. --tml)