Hack to improve tracing of Invokes when injecting in an existing process
In InjectDll, pass the name of the InitializeDll() function to the
NktHookLibHelpers::InjectDllByPidW() function (only). That means it
will be called only in the case when injecting in an already running
process. In that case we won't be able to catch Invoke calls of
IDispatch objects already instantiated, until a new such object that
uses the same implementation of Invoke (IDispatch_Invoke_Proxy in
oleaut32.dll, in the Word and Excel case at least) has been
instantiated.

Work around that by instantiating such an object ourselves there. This
works at least in a simple test case, but yeah, it does seem a bit
dangerous, so let's see. For now, use a "Word.Application" object, but
ideally we should use a dummy .exe file of our own. On the other hand,
if we do that we can't be sure it will actually use the same
implementation of IDispatch::Invoke.

Also fix the syntax of the Emacs mode line.